requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth-httplib2==0.1.0
google-auth==2.16.1
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
pip==22.3.1
protobuf==4.22.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==3.0.9
python-dotenv==0.21.1
requests==2.28.2
rsa==4.9
setuptools==66.1.1
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.14
wheel==0.38.4
